refactor: station controller, service 통합
admin, client 구분없이 하나의 클래스로 통합
